move swagger folder inside common folder
